# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FieldServicer Support Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets/{id}/attachments:
    get:
      operationId: getTicketAttachments
      tags: [attachments]
      summary: Get attachments for a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of attachments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
    post:
      operationId: uploadTicketAttachment
      tags: [attachments]
      summary: Upload a file attachment to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                uploadedBy:
                  type: string
      responses:
        "201":
          description: Created attachment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
  /attachments/{id}:
    delete:
      operationId: deleteAttachment
      tags: [attachments]
      summary: Delete an attachment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/count:
    get:
      operationId: getNotificationCount
      tags: [notifications]
      summary: Get unread notification count
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notification count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationCount"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /automation-rules:
    get:
      operationId: listAutomationRules
      tags: [automation]
      summary: List automation rules
      responses:
        "200":
          description: List of automation rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationRule"
    post:
      operationId: createAutomationRule
      tags: [automation]
      summary: Create an automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationRuleBody"
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
  /automation-rules/{id}:
    patch:
      operationId: updateAutomationRule
      tags: [automation]
      summary: Update an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationRuleBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
    delete:
      operationId: deleteAutomationRule
      tags: [automation]
      summary: Delete an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /simulate/email:
    post:
      operationId: simulateEmail
      tags: [simulate]
      summary: Simulate an inbound email creating a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateEmailBody"
      responses:
        "201":
          description: Created ticket from simulated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateResult"
  /simulate/whatsapp:
    post:
      operationId: simulateWhatsapp
      tags: [simulate]
      summary: Simulate an inbound WhatsApp message creating a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateWhatsappBody"
      responses:
        "201":
          description: Created ticket from simulated WhatsApp message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateResult"
components:
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
        ticketNumber:
          type: string
        subject:
          type: string
        description:
          type: string
        channel:
          type: string
          enum: [email, whatsapp, manual]
        status:
          type: string
          enum: [open, in_progress, waiting_customer, waiting_internal, escalated, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        siteId:
          type: integer
          nullable: true
        siteName:
          type: string
          nullable: true
        employeeId:
          type: integer
          nullable: true
        employeeName:
          type: string
          nullable: true
        jobId:
          type: integer
          nullable: true
        jobTitle:
          type: string
          nullable: true
        contactId:
          type: integer
          nullable: true
        contactName:
          type: string
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        slaBreached:
          type: boolean
        firstResponseAt:
          type: string
          format: date-time
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messageCount:
          type: integer
      required: [id, ticketNumber, subject, channel, status, priority, tags, slaBreached, createdAt, updatedAt, messageCount]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        ticketId:
          type: integer
          nullable: true
        type:
          type: string
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, type, title, body, read, createdAt]
    NotificationCount:
      type: object
      properties:
        count:
          type: integer
      required: [count]
    AutomationRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        triggerType:
          type: string
        conditions:
          type: array
          items: {}
        actions:
          type: array
          items: {}
        active:
          type: boolean
        runCount:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, triggerType, conditions, actions, active, runCount, createdAt, updatedAt]
    CreateAutomationRuleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        triggerType:
          type: string
        conditions:
          type: array
          items: {}
        actions:
          type: array
          items: {}
        active:
          type: boolean
      required: [name, triggerType]
    Attachment:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        fileName:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        storagePath:
          type: string
        uploadedBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, ticketId, fileName, originalName, mimeType, fileSize, storagePath, uploadedBy, createdAt]
...
    SimulateEmailBody:
      type: object
      properties:
        fromEmail:
          type: string
        fromName:
          type: string
        subject:
          type: string
        body:
          type: string
        priority:
          type: string
      required: [fromEmail, subject]
    SimulateWhatsappBody:
      type: object
      properties:
        fromPhone:
          type: string
        fromName:
          type: string
        message:
          type: string
        priority:
          type: string
      required: [fromPhone, message]
    SimulateResult:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/Ticket"
        autoLinkedClient:
          type: integer
          nullable: true
        autoLinkedContact:
          type: integer
          nullable: true
      required: [ticket]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]