# lib/api-zod/src/generated/types/simulateEmailBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FieldServicer Support Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulateEmailBody {
  fromEmail: string;
  fromName?: string;
  subject: string;
  body?: string;
  priority?: string;
}