# lib/api-zod/src/generated/types/simulateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FieldServicer Support Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Ticket } from "./ticket";

export interface SimulateResult {
  ticket: Ticket;
  autoLinkedClient?: number | null;
  autoLinkedContact?: number | null;
}