# lib/api-zod/src/generated/types/simulateWhatsappBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FieldServicer Support Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulateWhatsappBody {
  fromPhone: string;
  fromName?: string;
  message: string;
  priority?: string;
}