# lib/db/src/schema/notifications.ts
import { pgTable, serial, text, timestamp, integer, boolean, pgEnum } from "drizzle-orm/pg-core";
import { usersTable } from "./users";
import { ticketsTable } from "./tickets";

export const notificationTypeEnum = pgEnum("notification_type", [
  "ticket_assigned", "ticket_escalated", "sla_breach", "new_message",
  "ticket_created", "ticket_resolved", "automation_triggered", "mention"
]);

export const notificationsTable = pgTable("notifications", {
  id: serial("id").primaryKey(),
  userId: integer("user_id").references(() => usersTable.id),
  ticketId: integer("ticket_id").references(() => ticketsTable.id),
  type: notificationTypeEnum("type").notNull(),
  title: text("title").notNull(),
  body: text("body").notNull(),
  read: boolean("read").notNull().default(false),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export type Notification = typeof notificationsTable.$inferSelect;